Move DedicatedSessions into GdsSessions
