+++ src/graphdatascience/error/standalone_session_error.py
@@ -1,3 +1,3 @@
-class NotAvailableInSessions(Exception):
+class NotAvailableInStandaloneSessions(Exception):
     def __init__(self, subject: str) -> None:
         super().__init__(self, f"{subject} is not available in standalone sessions")
